refactor: rename Toaster component
