+++ src/components/index.js
@@ -11,4 +11,4 @@ export * from './calendar';
 export * from './pageSize';
 export * from './candidatesSearch';
 export * from './filterCandidates';
-export * from './toasterAlert';
+export * from './toaster';
