+++ src/components/toaster/index.js
@@ -1,0 +1,1 @@
+export { default as Toaster } from './Toaster';
