--- src/components/toasterAlert/index.js
@@ -1,1 +1,0 @@
-export { default as ToasterAlert } from './ToasterAlert';
